@@ -1,7 +1,7 @@
 package com.malinskiy.superrecyclerview.sample;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class GridActivity extends BaseActivity {
     @Override
