@@ -2,8 +2,8 @@ package com.malinskiy.superrecyclerview.sample;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.LinearLayoutManager;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import android.widget.Toast;
 
 import com.eowise.recyclerview.stickyheaders.StickyHeadersBuilder;
