@@ -1,6 +1,6 @@
 package com.malinskiy.superrecyclerview.sample;
 
-import android.support.v7.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
 
 public class ListActivity extends BaseActivity {
 
