@@ -1,7 +1,7 @@
 package com.malinskiy.superrecyclerview.sample;
 
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 public class StaggeredGridActivity extends BaseActivity {
     @Override
