@@ -1,6 +1,6 @@
 package com.malinskiy.superrecyclerview.swipe;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.malinskiy.superrecyclerview.R;
@@ -78,7 +78,7 @@ public abstract class BaseSwipeAdapter<VH extends BaseSwipeAdapter.BaseSwipeable
         public BaseSwipeableViewHolder(View itemView) {
             super(itemView);
 
-            swipeLayout = (SwipeLayout) itemView.findViewById(R.id.recyclerview_swipe);
+            swipeLayout = itemView.findViewById(R.id.recyclerview_swipe);
         }
     }
 }
