@@ -1,6 +1,6 @@
 package com.malinskiy.superrecyclerview.swipe;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 import java.util.Arrays;
